Safari-611.1.10/LayoutTests/fullscreen/fullscreen-user-agent-style-expected.txt

EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('width') == 'auto') OK
EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('height') == 'auto') OK
RUN(span.webkitRequestFullscreen())
EVENT(webkitfullscreenchange)
EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('width') == '100%') OK
EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('height') == '100%') OK
RUN(document.webkitExitFullscreen())
EVENT(webkitfullscreenchange)
EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('width') == 'auto') OK
EXPECTED (document.defaultView.getComputedStyle(span, null).getPropertyValue('height') == 'auto') OK
END OF TEST
